style(faq,svg-icon): move components into their own foler
