+++ components/svg-icon.vue
